--- a/scripts/ci/run_benchmarks.sh
+++ b/scripts/ci/run_benchmarks.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 set -e pipefail
 
 function run_benchmarks_for_folder {