# Issue

This PR adds CI job which allows to compare results of benchmarks in base branch (i.e. usually `master`) and PR. Should allow us to detect regressions and prove improvements. 

Works in the following way:
- Checkout and build base branch
- Checkout and build PR branch
- Run benchmarks from both branches
- Gather report with stdout of benchmarks from both base and PR and post it as footer section to PR description (or update if it already exists)

Example can be seen even in this PR, but it looks really ugly, because benchmarks from base branch currently contain a lot of not really relevant text which makes output too long. Example how it will look like after merge(after merge base branch will contain changes I made in benchmarks):

<img width="1263" alt="Screenshot 2024-05-11 at 22 26 58" src="https://github.com/Project-OSRM/osrm-backend/assets/266271/c2e213a6-c93d-4bad-82b6-97c03176bee4">

Also benchmarks themselves could be really improved and, may be, somehow unified, but I keep it out of scope of this PR.


## Tasklist

 - [ ] CHANGELOG.md entry ([How to write a changelog entry](http://keepachangelog.com/en/1.0.0/#how))
 - [ ] update relevant [Wiki pages](https://github.com/Project-OSRM/osrm-backend/wiki)
 - [ ] add tests (see [testing documentation](https://github.com/Project-OSRM/osrm-backend/blob/master/docs/testing.md))
 - [ ] review
 - [ ] adjust for comments
 - [ ] cherry pick to release branch

## Requirements / Relations

 Link any requirements here. Other pull requests this PR is based on?


<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | [2024-05-11T20:45:52.697546907] [info] aliased u32: 1141.83<br/>[2024-05-11T20:45:53.831361105] [info] plain u32: 1133.64<br/>[2024-05-11T20:45:55.010312885] [info] aliased double: 1178.9<br/>[2024-05-11T20:45:56.187701530] [info] plain double: 1177.34 | aliased u32: 1134.61<br/>plain u32: 1124.43<br/>aliased double: 1165.93<br/>plain double: 1178.22 |
| json-render | String: 8.50562ms<br/>Stringstream: 11.4028ms<br/>Vector: 7.75605ms | String: 8.46138ms<br/>Stringstream: 11.3649ms<br/>Vector: 7.6425ms |
| match_ch | 8.21244ms/req at 82 coordinate<br/>0.100152ms/coordinate | 8.24465ms/req at 82 coordinate<br/>0.100544ms/coordinate |
| match_mld | 7.46581ms/req at 82 coordinate<br/>0.0910464ms/coordinate | 7.10228ms/req at 82 coordinate<br/>0.0866132ms/coordinate |
| packedvector | [2024-05-11T20:48:02.991000738] [info] random write: std::vector 11174 ms, util::packed_vector 73866.4 ms. 6.61055<br/>[2024-05-11T20:48:02.991129828] [info] random read: std::vector 11089.8 ms, util::packed_vector 30574.1 ms. 2.75696 | random write:<br/>std::vector 11179.6 ms<br/>util::packed_vector 81950.3 ms<br/>slowdown: 7.33033<br/>random read:<br/>std::vector 11004 ms<br/>util::packed_vector 33840 ms<br/>slowdown: 3.07525 |
| rtree | Running raw RTree queries (1 result) with 10000 coordinates: Took 0.206803 seconds (206.804ms)  ->  0.0206804 ms/query (206.804ms)<br/>Running raw RTree queries (10 results) with 10000 coordinates: Took 0.242288 seconds (242.288ms)  ->  0.0242288 ms/query (242.288ms) | 1 result:<br/>208.412ms ->  0.0208412 ms/query<br/>10 results:<br/>243.502ms ->  0.0243502 ms/query |
<!-- BENCHMARK_RESULTS_END -->