@DennisOSRM 2024-05-14
Love it. Is there any chance these metrics could be published as a time series somewhere? Seeing trends over time might be super insightful.