@SiarheiFedartsou 2024-05-14
> Love it. Is there any chance these metrics could be published as a time series somewhere? Seeing trends over time might be super insightful.

It is definitely somehow possible, but not sure how much efforts we will need: AFAIK there is nothing like this in GitHub infra, so we would need something deployed somewhere - probably a bit of overkill for now. But I'll think if it is doable with some low efforts :+1: